requirements.txt
requests
transformers
datasets
faiss-cpu
torch
